refactor(frontend): extract account settings form primitives
Move the reusable account-settings form toolkit out of the page: the
non-component helpers (actionButtonSx, SectionSubmit, useSectionSubmit) into
accountSettingsForm.ts and the presentational components (SectionAlerts,
InlineEditor, SettingsCard) into accountSettingsCards.tsx. Splitting
components from the hook/const keeps each file fast-refresh clean. The page
drops the imports only those pieces used. No behavior change.
